backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
openai
PyPDF2
numpy
pgvector
python-multipart
cloudinary
requests
pydub
groq
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator
python-socketio
google-auth
pydantic-settings